+++ test/core/implicit/id_token.authorization.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const {
-  provider, agent, AuthenticationRequest, getSession, wrap
+  provider, agent, AuthorizationRequest, getSession, wrap
 } = require('../../test_helper')(__dirname);
 const sinon = require('sinon');
 const { expect } = require('chai');
@@ -18,7 +18,7 @@ provider.setupCerts();
     before(agent.login);
 
     it('responds with a id_token in fragment', function () {
-      const auth = new AuthenticationRequest({
+      const auth = new AuthorizationRequest({
         response_type: 'id_token',
         scope: 'openid'
       });
@@ -41,7 +41,7 @@ provider.setupCerts();
       delete session.loginTs;
       delete session.account;
 
-      const auth = new AuthenticationRequest({
+      const auth = new AuthorizationRequest({
         response_type: 'id_token',
         scope: 'openid'
       });
@@ -54,7 +54,7 @@ provider.setupCerts();
 
     describe('requested by the End-User', function () {
       it('login was requested by the client by prompt parameter', function () {
-        const auth = new AuthenticationRequest({
+        const auth = new AuthorizationRequest({
           response_type: 'id_token',
           prompt: 'login',
           scope: 'openid'
@@ -66,11 +66,11 @@ provider.setupCerts();
         .expect(auth.validateInteractionError('login_required', 'login_prompt'));
       });
 
-      it('session is too old for this authentication request', function () {
+      it('session is too old for this authorization request', function () {
         const session = getSession(agent);
         session.loginTs = (new Date() / 1000 | 0) - 3600; // an hour ago
 
-        const auth = new AuthenticationRequest({
+        const auth = new AuthorizationRequest({
           response_type: 'id_token',
           max_age: '1800', // 30 minutes old session max
           scope: 'openid'
@@ -90,8 +90,8 @@ provider.setupCerts();
     it('dupe parameters', function () {
     // fake a query like this scope=openid&scope=openid
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
-      const auth = new AuthenticationRequest({
+      provider.once('authorization.error', spy);
+      const auth = new AuthorizationRequest({
         response_type: 'id_token',
         scope: ['openid', 'openid']
       });
@@ -110,8 +110,8 @@ provider.setupCerts();
 
     it('disallowed response mode', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
-      const auth = new AuthenticationRequest({
+      provider.once('authorization.error', spy);
+      const auth = new AuthorizationRequest({
         response_type: 'code id_token',
         scope: 'openid',
         response_mode: 'query'
@@ -132,8 +132,8 @@ provider.setupCerts();
     ['request', 'request_uri', 'registration'].forEach(function (param) {
       it(`not supported parameter ${param}`, function () {
         const spy = sinon.spy();
-        provider.once('authentication.error', spy);
-        const auth = new AuthenticationRequest({
+        provider.once('authorization.error', spy);
+        const auth = new AuthorizationRequest({
           response_type: 'id_token',
           scope: 'openid',
           [param]: 'some'
@@ -165,8 +165,8 @@ provider.setupCerts();
       it('missing mandatory parameter redirect_uri', function () {
         const emitSpy = sinon.spy();
         const renderSpy = sinon.spy(provider.configuration(), 'renderError');
-        provider.once('authentication.error', emitSpy);
-        const auth = new AuthenticationRequest({
+        provider.once('authorization.error', emitSpy);
+        const auth = new AuthorizationRequest({
           response_type: 'id_token',
           scope: 'openid'
         });
@@ -191,8 +191,8 @@ provider.setupCerts();
     ['client_id', 'scope', 'nonce'].forEach(function (param) {
       it(`missing mandatory parameter ${param}`, function () {
         const spy = sinon.spy();
-        provider.once('authentication.error', spy);
-        const auth = new AuthenticationRequest({
+        provider.once('authorization.error', spy);
+        const auth = new AuthorizationRequest({
           response_type: 'id_token',
           scope: 'openid'
         });
@@ -215,8 +215,8 @@ provider.setupCerts();
 
     it('missing mandatory parameter response_type', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
-      const auth = new AuthenticationRequest({
+      provider.once('authorization.error', spy);
+      const auth = new AuthorizationRequest({
         response_type: 'id_token',
         scope: 'openid'
       });
@@ -237,8 +237,8 @@ provider.setupCerts();
 
     it('unsupported prompt', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
-      const auth = new AuthenticationRequest({
+      provider.once('authorization.error', spy);
+      const auth = new AuthorizationRequest({
         response_type: 'id_token',
         scope: 'openid',
         prompt: 'unsupported'
@@ -259,8 +259,8 @@ provider.setupCerts();
 
     it('bad prompt combination', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
-      const auth = new AuthenticationRequest({
+      provider.once('authorization.error', spy);
+      const auth = new AuthorizationRequest({
         response_type: 'id_token',
         scope: 'openid',
         prompt: 'none login'
@@ -281,8 +281,8 @@ provider.setupCerts();
 
     it('unsupported scope', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
-      const auth = new AuthenticationRequest({
+      provider.once('authorization.error', spy);
+      const auth = new AuthorizationRequest({
         response_type: 'id_token',
         scope: 'openid and unsupported'
       });
@@ -302,8 +302,8 @@ provider.setupCerts();
 
     it('missing openid scope', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
-      const auth = new AuthenticationRequest({
+      provider.once('authorization.error', spy);
+      const auth = new AuthorizationRequest({
         response_type: 'id_token',
         scope: 'profile'
       });
@@ -323,8 +323,8 @@ provider.setupCerts();
 
     it('invalid use of scope offline_access', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
-      const auth = new AuthenticationRequest({
+      provider.once('authorization.error', spy);
+      const auth = new AuthorizationRequest({
         response_type: 'id_token',
         scope: 'openid offline_access'
       });
@@ -344,8 +344,8 @@ provider.setupCerts();
 
     it('unrecognized client_id provided', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
-      const auth = new AuthenticationRequest({
+      provider.once('authorization.error', spy);
+      const auth = new AuthorizationRequest({
         response_type: 'id_token',
         scope: 'openid',
         client_id: 'unrecognized'
@@ -366,8 +366,8 @@ provider.setupCerts();
 
     it('unsupported response_type', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
-      const auth = new AuthenticationRequest({
+      provider.once('authorization.error', spy);
+      const auth = new AuthorizationRequest({
         response_type: 'unsupported',
         scope: 'openid'
       });
@@ -386,8 +386,8 @@ provider.setupCerts();
 
     it('restricted response_type', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
-      const auth = new AuthenticationRequest({
+      provider.once('authorization.error', spy);
+      const auth = new AuthorizationRequest({
         response_type: 'code',
         scope: 'openid'
       });
@@ -407,8 +407,8 @@ provider.setupCerts();
     it('redirect_uri mismatch', function () {
       const emitSpy = sinon.spy();
       const renderSpy = sinon.spy(provider.configuration(), 'renderError');
-      provider.once('authentication.error', emitSpy);
-      const auth = new AuthenticationRequest({
+      provider.once('authorization.error', emitSpy);
+      const auth = new AuthorizationRequest({
         response_type: 'id_token',
         scope: 'openid',
         redirect_uri: 'http://example.client.dev/notregistered'
@@ -434,8 +434,8 @@ provider.setupCerts();
 
       it('malformed id_token_hint', function () {
         const spy = sinon.spy();
-        provider.once('authentication.error', spy);
-        const auth = new AuthenticationRequest({
+        provider.once('authorization.error', spy);
+        const auth = new AuthorizationRequest({
           response_type: 'id_token',
           scope: 'openid',
           id_token_hint: 'invalid'
@@ -465,7 +465,7 @@ provider.setupCerts();
       });
 
       it('responds with server_error redirect to redirect_uri', function () {
-        const auth = new AuthenticationRequest({
+        const auth = new AuthorizationRequest({
           response_type: 'id_token',
           prompt: 'none',
           scope: 'openid'