+++ test/core/userinfo/userinfo.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const {
-  provider, agent, AuthenticationRequest, wrap
+  provider, agent, AuthorizationRequest, wrap
 } = require('../../test_helper')(__dirname);
 const { expect } = require('chai');
 const url = require('url');
@@ -13,7 +13,7 @@ describe('userinfo /me', function () {
   before(agent.login);
 
   before(function () {
-    const auth = new AuthenticationRequest({
+    const auth = new AuthorizationRequest({
       response_type: 'id_token token',
       scope: 'openid email'
     });