+++ test/request/uri_request.test.js
@@ -169,7 +169,7 @@ describe('configuration features.requestUri', function () {
 
     it('doesnt allow too long request_uris', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
+      provider.once('authorization.error', spy);
 
       return wrap({
         agent,
@@ -268,7 +268,7 @@ describe('configuration features.requestUri', function () {
 
       it('doesnt allow to bypass these', function () {
         const spy = sinon.spy();
-        provider.once('authentication.error', spy);
+        provider.once('authorization.error', spy);
 
         return wrap({
           agent,
@@ -293,7 +293,7 @@ describe('configuration features.requestUri', function () {
 
     it('doesnt allow slow requests (socket delay)', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
+      provider.once('authorization.error', spy);
 
       nock('https://client.example.com')
         .get('/request')
@@ -321,7 +321,7 @@ describe('configuration features.requestUri', function () {
 
     it('doesnt allow slow requests (response delay)', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
+      provider.once('authorization.error', spy);
 
       nock('https://client.example.com')
         .get('/request')
@@ -349,7 +349,7 @@ describe('configuration features.requestUri', function () {
 
     it('doesnt accepts 200s, rejects even on redirect', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
+      provider.once('authorization.error', spy);
 
       nock('https://client.example.com')
         .get('/request')
@@ -378,7 +378,7 @@ describe('configuration features.requestUri', function () {
 
     it('doesnt allow request inception', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
+      provider.once('authorization.error', spy);
 
       return JWT.sign({
         client_id: 'client',
@@ -412,7 +412,7 @@ describe('configuration features.requestUri', function () {
 
     it('doesnt allow requestUri inception', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
+      provider.once('authorization.error', spy);
 
       return JWT.sign({
         client_id: 'client',
@@ -447,7 +447,7 @@ describe('configuration features.requestUri', function () {
 
     it('doesnt allow response_type to differ', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
+      provider.once('authorization.error', spy);
 
       return JWT.sign({
         client_id: 'client',
@@ -481,7 +481,7 @@ describe('configuration features.requestUri', function () {
 
     it('doesnt allow client_id to differ', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
+      provider.once('authorization.error', spy);
 
       return JWT.sign({
         client_id: 'client2',
@@ -515,7 +515,7 @@ describe('configuration features.requestUri', function () {
 
     it('handles invalid signed looklike jwts', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
+      provider.once('authorization.error', spy);
 
       nock('https://client.example.com')
         .get('/request')
@@ -544,7 +544,7 @@ describe('configuration features.requestUri', function () {
 
     it('doesnt allow clients with predefined alg to bypass this alg', function () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
+      provider.once('authorization.error', spy);
 
       return JWT.sign({
         client_id: 'client-with-HS-sig',
@@ -579,7 +579,7 @@ describe('configuration features.requestUri', function () {
 
     it('bad signatures will be rejected', function * () {
       const spy = sinon.spy();
-      provider.once('authentication.error', spy);
+      provider.once('authorization.error', spy);
 
       const key = (yield Client.find('client-with-HS-sig')).keystore.get('clientSecret');
       return JWT.sign({