+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Following semver, 1.0.0 will mark the first API stable release and commence of this file,
 until then please use the compare views of github for reference.
 
+- https://github.com/panva/node-oidc-provider/compare/v0.6.0...v0.7.0
+  - all things `authentication` renamed to `authorization`
 - https://github.com/panva/node-oidc-provider/compare/v0.5.0...v0.6.0
 - https://github.com/panva/node-oidc-provider/compare/v0.4.0...v0.5.0
 - https://github.com/panva/node-oidc-provider/compare/v0.3.1...v0.4.0